(#143) implement mute / ban / blocked
[WS:Channel]
- add demote, unban, ban
- add check for everything in the channel
- display banned user on channel management
- display demote on non admin and vis versa
- remove muted table
- add / remove private mode
- add mute for N sec / add unmute

[WS:blocked]
- add block/unblock button/func
- rename websocket file
- blocked user doesn't appear in dm / channel chat anymore

+++ src/nestjs/src/modules/database/notification/service.ts
@@ -18,15 +18,13 @@ export class DBNotificationService {
 		notif.userId = post.userId;
 		notif.data = post.data;
 		const ret = await this.NotificationRepo.save(notif);
-		return (ret);
+		return ret;
 	}
 
 	async update(id: number, post: DBNotificationPost) {
-		const tmp = await this.NotificationRepo.findOneBy({id: id});
-		if (tmp)
-			await this.NotificationRepo.update(id, post);
-		else
-			throw new NotFoundException("Notification not found");
+		const tmp = await this.NotificationRepo.findOneBy({ id: id });
+		if (tmp) await this.NotificationRepo.update(id, post);
+		else throw new NotFoundException("Notification not found");
 	}
 
 	async returnAll() {
@@ -34,26 +32,21 @@ export class DBNotificationService {
 	}
 
 	async returnOne(id: number) {
-		const tmp = await this.NotificationRepo.findOneBy({id: id});
-		if (tmp)
-			return (tmp);
-		else
-			throw new NotFoundException("Notification not found");
+		const tmp = await this.NotificationRepo.findOneBy({ id: id });
+		if (tmp) return tmp;
+		else throw new NotFoundException("Notification not found");
 	}
 
 	async delete(id: number) {
-		const tmp = await this.NotificationRepo.findOneBy({id: id});
-		if (tmp)
-			return await this.NotificationRepo.delete(tmp.id);
-		else
-			throw new NotFoundException("Notification not found");
+		const tmp = await this.NotificationRepo.findOneBy({ id: id });
+		if (tmp) return await this.NotificationRepo.delete(tmp.id);
+		else throw new NotFoundException("Notification not found");
 	}
 
-	async getNotif(user_id: number, data?: string, type?: NotificationType)
-	{
+	async getNotif(user_id: number, data?: string, type?: NotificationType) {
 		var where_query: any = {
 			userId: user_id,
-		}
+		};
 		if (data) where_query["data"] = data;
 		if (type) where_query["type"] = type;
 		return await this.NotificationRepo.findOne({
@@ -64,8 +57,7 @@ export class DBNotificationService {
 		});
 	}
 
-	async getNotifByUserId(id: number)
-	{
+	async getNotifByUserId(id: number) {
 		return await this.NotificationRepo.find({
 			where: {
 				userId: id,
